impl: fix "view source" link
The "View source on GitHub" link in the footer broke in the switch to
Netlify due to the lack of `site.github.source`. (They have always been
broken on Netlify deploy previews.)

To fix:

- Hard-code the source path (`/docs`) since this is static.
- Link to the specific commit rather than the branch. This is actually
  better for the user because it is the precise source code that was
  used to build, rather than whatever is on the latest branch.

While we're at it, hard-code the github repo link in the header, since
that is also static. This ways if the metadata breaks in the future, the
site will fail gracefully and only the "view source" link will be
impacted.

Finally, append `?plain=1` to the URL so that we link to the actual
Markdown source rather than a rendered preview.

Fixes #922.

Signed-off-by: Mark Lodato <lodato@google.com>